# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: RedisShake Control Plane API
  version: 0.1.0
  description: API contract for the RedisShake Web control plane.
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Process liveness
      responses:
        "200":
          description: The control-plane process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /readyz:
    get:
      operationId: getReadiness
      summary: Storage readiness
      responses:
        "200":
          description: The control plane and SQLite storage are ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Readiness"
        "503":
          description: SQLite storage is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/system/info:
    get:
      operationId: getSystemInfo
      summary: Control-plane build and storage information
      responses:
        "200":
          description: System information without credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfo"
  /api/v1/connections:
    get:
      operationId: listConnections
      summary: List saved Redis connections
      responses:
        "200":
          description: Saved connections without credential material.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      summary: Create a Redis connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionInput"
      responses:
        "201":
          description: Connection created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          description: A master key is required before storing credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/connections/{id}:
    parameters:
      - $ref: "#/components/parameters/ConnectionID"
    get:
      operationId: getConnection
      summary: Get a saved Redis connection
      responses:
        "200":
          description: Connection without credential material.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateConnection
      summary: Update selected Redis connection fields
      description: Omitted credential fields are preserved. Send an empty credential value to clear it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionPatch"
      responses:
        "200":
          description: Updated connection without credential material.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteConnection
      summary: Delete an unused Redis connection
      responses:
        "204":
          description: Connection deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The connection is referenced by a task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/connections/test:
    post:
      operationId: testUnsavedConnection
      summary: Test connection settings without saving them
      description: A target test writes a random temporary Key with a 60-second TTL and then deletes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection, purpose]
              properties:
                connection:
                  $ref: "#/components/schemas/ConnectionInput"
                purpose:
                  $ref: "#/components/schemas/TestPurpose"
      responses:
        "200":
          description: Structured test result. Connection failures are represented by success=false.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/connections/{id}/test:
    parameters:
      - $ref: "#/components/parameters/ConnectionID"
    post:
      operationId: testSavedConnection
      summary: Test a saved Redis connection
      description: A target test writes a random temporary Key with a 60-second TTL and then deletes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose]
              properties:
                purpose:
                  $ref: "#/components/schemas/TestPurpose"
      responses:
        "200":
          description: Structured test result persisted as the connection's latest result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/connections/{id}/copy:
    parameters:
      - $ref: "#/components/parameters/ConnectionID"
    post:
      operationId: copyConnection
      summary: Copy a connection and re-encrypt its stored credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyRequest"
      responses:
        "201":
          description: Connection copy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tasks:
    get:
      operationId: listTasks
      summary: List synchronization tasks
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Synchronization tasks.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      summary: Create a task draft
      description: Source and target may be omitted while the creation wizard is incomplete.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Draft created at config revision 1.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/TaskID"
    get:
      operationId: getTask
      summary: Get a synchronization task
      responses:
        "200":
          description: Task and latest precheck result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      summary: Update a task draft or ready task
      description: A successful update increments config_revision, returns the task to DRAFT, and invalidates the previous precheck.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Updated draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Revision conflict or archived task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: archiveTask
      summary: Archive a task without deleting history
      responses:
        "204":
          description: Task archived.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The task has an active run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tasks/{id}/precheck:
    parameters:
      - $ref: "#/components/parameters/TaskID"
    post:
      operationId: precheckTask
      summary: Run source, target, topology, configuration, and danger checks
      description: The target check writes and deletes a random Key with a 60-second TTL. The task becomes READY only when all checks pass and warnings are acknowledged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrecheckRequest"
      responses:
        "200":
          description: Structured precheck result. Validation failures are represented by ready=false.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrecheckResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Revision conflict or archived task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tasks/{id}/copy:
    parameters:
      - $ref: "#/components/parameters/TaskID"
    post:
      operationId: copyTask
      summary: Copy a task as a new DRAFT at revision 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyRequest"
      responses:
        "201":
          description: Task copy requiring a new precheck.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tasks/{id}/runs:
    parameters:
      - $ref: "#/components/parameters/TaskID"
    get:
      operationId: listTaskRuns
      summary: List task run history
      responses:
        "200":
          description: Runs ordered by newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
    post:
      operationId: startTaskRun
      summary: Start one RedisShake worker for a READY task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "201":
          description: Worker started and its loopback status endpoint became readable, or a scan completed before the handshake finished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "409":
          description: Task is not READY, revision changed, or an active/unknown run already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: A Run record was created but the worker failed during startup.
          content:
            application/json:
              schema:
                type: object
                required: [error, run]
                properties:
                  error:
                    $ref: "#/components/schemas/Error"
                  run:
                    $ref: "#/components/schemas/Run"
        "503":
          description: RedisShake worker binary or version probe is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/runs/{id}:
    parameters:
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: getRun
      summary: Get persisted process and RedisShake status
      responses:
        "200":
          description: Run state and latest status payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{id}/stop:
    parameters:
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: stopRun
      summary: Request graceful process-group termination
      responses:
        "202":
          description: Run entered STOPPING.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "409":
          description: Run is terminal or ownership cannot be proven after restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/runs/{id}/force-stop:
    parameters:
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: forceStopRun
      summary: Force process-group termination after user confirmation
      responses:
        "202":
          description: Force termination was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "409":
          description: Run is terminal or ownership cannot be proven after restart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/runs/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/RunID"
      - name: offset
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 1048576
          default: 65536
    get:
      operationId: getRunLogs
      summary: Read a redacted page of worker stdout/stderr
      responses:
        "200":
          description: Log content and next byte offset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogResult"
  /api/v1/runs/{id}/events:
    parameters:
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: streamRunEvents
      summary: Stream persisted run status updates
      responses:
        "200":
          description: Server-sent status events until the run reaches a terminal state.
          content:
            text/event-stream:
              schema:
                type: string
components:
  parameters:
    ConnectionID:
      name: id
      in: path
      required: true
      schema:
        type: string
    TaskID:
      name: id
      in: path
      required: true
      schema:
        type: string
    RunID:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid JSON or connection fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request conflicts with existing data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [status, started_at]
      properties:
        status:
          const: ok
        started_at:
          type: string
          format: date-time
    Readiness:
      type: object
      required: [status, database]
      properties:
        status:
          const: ready
        database:
          const: ok
    SystemInfo:
      type: object
      required: [version, git_commit, storage, data_dir, runtime_dir, secrets_configured, worker_path, web_ui_configured, max_concurrent_runs, log_retention_days]
      properties:
        version:
          type: string
        git_commit:
          type: string
        storage:
          const: sqlite
        data_dir:
          type: string
        runtime_dir:
          type: string
        secrets_configured:
          type: boolean
        worker_path:
          type: string
        web_ui_configured:
          type: boolean
        max_concurrent_runs:
          type: integer
          minimum: 1
        log_retention_days:
          type: integer
          minimum: 0
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    CopyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 128
    TLSInput:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        server_name:
          type: string
        insecure_skip_verify:
          type: boolean
          default: false
        ca_cert_pem:
          type: string
          writeOnly: true
        client_cert_pem:
          type: string
          writeOnly: true
        client_key_pem:
          type: string
          writeOnly: true
    SentinelInput:
      type: object
      properties:
        address:
          type: string
          examples: ["127.0.0.1:26379"]
        master_name:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        tls:
          $ref: "#/components/schemas/TLSInput"
    ConnectionInput:
      type: object
      required: [name, topology]
      properties:
        name:
          type: string
          maxLength: 128
        topology:
          type: string
          enum: [standalone, sentinel, cluster]
        address:
          type: string
          description: Redis host:port. Optional for Sentinel because the master is resolved dynamically.
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        tls:
          $ref: "#/components/schemas/TLSInput"
        sentinel:
          $ref: "#/components/schemas/SentinelInput"
    TLSPatch:
      allOf:
        - $ref: "#/components/schemas/TLSInput"
      description: Only supplied properties are changed. Disabling TLS clears stored TLS material.
    SentinelPatch:
      allOf:
        - $ref: "#/components/schemas/SentinelInput"
      description: Only supplied properties are changed.
    ConnectionPatch:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
        topology:
          type: string
          enum: [standalone, sentinel, cluster]
        address:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        tls:
          $ref: "#/components/schemas/TLSPatch"
        sentinel:
          $ref: "#/components/schemas/SentinelPatch"
    TLSView:
      type: object
      required: [enabled, insecure_skip_verify, ca_cert_configured, client_cert_configured, client_key_configured]
      properties:
        enabled:
          type: boolean
        server_name:
          type: string
        insecure_skip_verify:
          type: boolean
        ca_cert_configured:
          type: boolean
        client_cert_configured:
          type: boolean
        client_key_configured:
          type: boolean
    SentinelView:
      type: object
      required: [password_configured, tls]
      properties:
        address:
          type: string
        master_name:
          type: string
        username:
          type: string
        password_configured:
          type: boolean
        tls:
          $ref: "#/components/schemas/TLSView"
    Connection:
      type: object
      required: [id, name, topology, password_configured, tls, sentinel, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        topology:
          type: string
          enum: [standalone, sentinel, cluster]
        address:
          type: string
        username:
          type: string
        password_configured:
          type: boolean
        tls:
          $ref: "#/components/schemas/TLSView"
        sentinel:
          $ref: "#/components/schemas/SentinelView"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_tested_at:
          type: [string, "null"]
          format: date-time
        last_test_result:
          oneOf:
            - $ref: "#/components/schemas/ConnectionTestResult"
            - type: "null"
    TestPurpose:
      type: string
      enum: [source, target]
    CheckItem:
      type: object
      required: [code, state, message]
      properties:
        code:
          type: string
        state:
          type: string
          enum: [PASS, WARNING, FAIL]
        message:
          type: string
    ConnectionTestResult:
      type: object
      required: [success, purpose, cluster_enabled, latency_ms, checks, tested_at]
      properties:
        success:
          type: boolean
        purpose:
          $ref: "#/components/schemas/TestPurpose"
        effective_address:
          type: string
        server_product:
          type: string
        server_version:
          type: string
        role:
          type: string
        cluster_enabled:
          type: boolean
        latency_ms:
          type: integer
          format: int64
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckItem"
        tested_at:
          type: string
          format: date-time
    SyncReaderOptions:
      type: object
      required: [sync_rdb, sync_aof, prefer_replica, try_diskless]
      properties:
        sync_rdb:
          type: boolean
        sync_aof:
          type: boolean
        prefer_replica:
          type: boolean
        try_diskless:
          type: boolean
    ScanReaderOptions:
      type: object
      required: [dbs, scan, ksn, count, prefer_replica, skip_unknown_type]
      properties:
        dbs:
          type: array
          items:
            type: integer
        scan:
          type: boolean
        ksn:
          type: boolean
        count:
          type: integer
          minimum: 1
        prefer_replica:
          type: boolean
        skip_unknown_type:
          type: array
          items:
            type: string
    FilterOptions:
      type: object
      properties:
        allow_keys: {type: array, items: {type: string}}
        allow_key_prefix: {type: array, items: {type: string}}
        allow_key_suffix: {type: array, items: {type: string}}
        allow_key_regex: {type: array, items: {type: string}}
        block_keys: {type: array, items: {type: string}}
        block_key_prefix: {type: array, items: {type: string}}
        block_key_suffix: {type: array, items: {type: string}}
        block_key_regex: {type: array, items: {type: string}}
        allow_db: {type: array, items: {type: integer}}
        block_db: {type: array, items: {type: integer}}
        allow_command: {type: array, items: {type: string}}
        block_command: {type: array, items: {type: string}}
        allow_command_group: {type: array, items: {type: string}}
        block_command_group: {type: array, items: {type: string}}
        function: {type: string}
    AdvancedOptions:
      type: object
      properties:
        rdb_restore_command_behavior:
          type: string
          enum: [panic, rewrite, skip]
        pipeline_count_limit:
          type: integer
          minimum: 1
        target_redis_max_qps:
          type: integer
          minimum: 1
          maximum: 300000
        target_redis_client_max_querybuf_len:
          type: integer
          minimum: 1
        target_redis_proto_max_bulk_len:
          type: integer
          minimum: 1048576
        empty_db_before_sync:
          type: boolean
    TaskSpec:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
        description:
          type: string
        mode:
          type: string
          enum: [sync, scan]
        source_connection_id:
          type: string
        target_connection_id:
          type: string
        sync_reader:
          $ref: "#/components/schemas/SyncReaderOptions"
        scan_reader:
          $ref: "#/components/schemas/ScanReaderOptions"
        filter:
          $ref: "#/components/schemas/FilterOptions"
        advanced:
          $ref: "#/components/schemas/AdvancedOptions"
    TaskCreate:
      allOf:
        - type: object
          required: [name, mode]
        - $ref: "#/components/schemas/TaskSpec"
    TaskPatch:
      allOf:
        - type: object
          required: [expected_revision]
          properties:
            expected_revision:
              type: integer
              minimum: 1
        - $ref: "#/components/schemas/TaskSpec"
      description: Only supplied TaskSpec fields are changed.
    Task:
      type: object
      required: [id, spec, state, config_revision, created_at, updated_at]
      properties:
        id:
          type: string
        spec:
          $ref: "#/components/schemas/TaskSpec"
        state:
          type: string
          enum: [DRAFT, READY, ARCHIVED]
        config_revision:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_prechecked_at:
          type: [string, "null"]
          format: date-time
        last_precheck_result:
          oneOf:
            - $ref: "#/components/schemas/PrecheckResult"
            - type: "null"
    PrecheckRequest:
      type: object
      required: [expected_revision]
      properties:
        expected_revision:
          type: integer
          minimum: 1
        acknowledge_warnings:
          type: boolean
          default: false
    PrecheckResult:
      type: object
      required: [task_id, config_revision, ready, checks, checked_at]
      properties:
        task_id:
          type: string
        config_revision:
          type: integer
        ready:
          type: boolean
        config_digest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckItem"
        checked_at:
          type: string
          format: date-time
    StartRunRequest:
      type: object
      required: [expected_revision]
      properties:
        expected_revision:
          type: integer
          minimum: 1
    Run:
      type: object
      required: [id, task_id, config_revision, config_snapshot, state, started_at, stop_requested_by_user, status_healthy, updated_at]
      properties:
        id:
          type: string
        task_id:
          type: string
        config_revision:
          type: integer
        config_snapshot:
          $ref: "#/components/schemas/TaskSpec"
        state:
          type: string
          enum: [STARTING, RUNNING, STOPPING, STOPPED, SUCCEEDED, FAILED, UNKNOWN]
        pid:
          type: integer
        status_port:
          type: integer
        started_at:
          type: string
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
        exit_code:
          type: [integer, "null"]
        exit_reason:
          type: string
        last_heartbeat_at:
          type: [string, "null"]
          format: date-time
        stop_requested_by_user:
          type: boolean
        status_healthy:
          type: boolean
        status:
          type: object
          additionalProperties: true
        updated_at:
          type: string
          format: date-time
    LogResult:
      type: object
      required: [content, next_offset, eof]
      properties:
        content:
          type: string
        next_offset:
          type: integer
          format: int64
        eof:
          type: boolean